requirements.txt
python-telegram-bot[rate-limiter,webhooks]~=20.0
boto3~=1.26
celery~=5.2
redis~=4.4
requests~=2.28
bjoern~=3.2
